+++ tests/testthat/test-sbp.R
@@ -9,6 +9,11 @@ colnames(sbp2) <- NULL
 test_that("sbp.fromHclust matches the ape and philr method", {
 
   expect_equal(
+    sbp1[,1],
+    sbp2[,1]
+  )
+
+  expect_equal(
     sbp.sort(sbp1),
     sbp.sort(sbp2)
   )